Move client's views directory
