+++ client/javascript/components/views/Login.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import ApplicationView from '../components/layout/ApplicationView';
-import RegistrationForm from '../components/auth/RegistrationForm';
+import ApplicationView from '../layout/ApplicationView';
+import LoginForm from '../auth/LoginForm';
 
 export default class LoginView extends React.Component {
   render() {
     return (
       <ApplicationView modifier="auth-form">
-        <RegistrationForm />
+        <LoginForm />
       </ApplicationView>
     );
   }
