+++ client/javascript/components/views/Register.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import ApplicationView from '../components/layout/ApplicationView';
-import LoginForm from '../components/auth/LoginForm';
+import ApplicationView from '../layout/ApplicationView';
+import RegistrationForm from '../auth/RegistrationForm';
 
 export default class LoginView extends React.Component {
   render() {
     return (
       <ApplicationView modifier="auth-form">
-        <LoginForm />
+        <RegistrationForm />
       </ApplicationView>
     );
   }
