+++ client/javascript/components/views/TorrentList.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import Alerts from '../components/alerts/Alerts';
-import ApplicationView from '../components/layout/ApplicationView';
-import ApplicationContent from '../components/layout/ApplicationContent';
-import Modals from '../components/modals/Modals';
-import Sidebar from '../components/panels/Sidebar';
-import SettingsStore from '../stores/SettingsStore';
-import TorrentActions from '../actions/TorrentActions';
-import TorrentList from '../components/panels/TorrentList';
+import Alerts from '../alerts/Alerts';
+import ApplicationView from '../layout/ApplicationView';
+import ApplicationContent from '../layout/ApplicationContent';
+import Modals from '../modals/Modals';
+import Sidebar from '../panels/Sidebar';
+import SettingsStore from '../../stores/SettingsStore';
+import TorrentActions from '../../actions/TorrentActions';
+import TorrentList from '../panels/TorrentList';
 
 export default class TorrentListView extends React.Component {
   componentDidMount() {
